Combined head and eye face tracking movement into one node
