--- src/eye_movement/setup.cfg
@@ -1,4 +1,0 @@
-[develop]
-script-dir=$base/lib/eye_movement
-[install]
-install-scripts=$base/lib/eye_movement
