+++ src/face_tracker_movement/setup.cfg
@@ -1,0 +1,4 @@
+[develop]
+script-dir=$base/lib/face_tracker_movement
+[install]
+install-scripts=$base/lib/face_tracker_movement
